+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ v3.8.1-rc2 (UNRELEASED)
 
 * :+1: Added `set_associations` and `remove_associations` method to the `Widget` class (#827)
 * :+1: Implemented create, edit and delete functionality for `Associations` (#827)
+* :+1: Added `association()` method to the `Client` class, to retrieve a single association.
 
 v3.8.1-rc1 (14AUG20)
 --------------------